# openapi.yaml
openapi: '3.1.0'
info:
  title: Photo Gallery API
  description: |
    A RESTful API for managing a personal photo gallery.
    
    ## Authentication
    All API endpoints require an API key passed as a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-api-key>
    ```
    
    Generate an API key using the CLI script:
    ```
    node scripts/generate-api-key.js --label "My App"
    ```
    
    ## Rate Limiting
    API requests are rate-limited to **100 requests per minute** per API key.
    Rate limit information is returned in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
    - `RateLimit-Remaining`: Remaining requests in the current window
    - `RateLimit-Reset`: Time when the rate limit window resets
    
  version: '1.0.0'
  contact:
    name: Photo Gallery API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key generated via `node scripts/generate-api-key.js`
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: integer
          description: Unique image identifier
          example: 42
        originalName:
          type: string
          description: Original filename of the uploaded image
          example: "sunset-beach.jpg"
        filename:
          type: string
          description: Stored filename on disk
          example: "1720000000000-123456789.jpg"
        mimeType:
          type: string
          description: MIME type of the image
          example: "image/jpeg"
        fileSize:
          type: integer
          description: File size in bytes
          example: 2048576
        width:
          type: integer
          description: Image width in pixels
          example: 3840
        height:
          type: integer
          description: Image height in pixels
          example: 2160
        title:
          type: string
          nullable: true
          description: Optional title
          example: "Sunset at the beach"
        description:
          type: string
          nullable: true
          description: Optional description
          example: "A beautiful sunset photo"
        altText:
          type: string
          nullable: true
          description: Alt text for accessibility
          example: "Orange sunset over the ocean"
        cameraMake:
          type: string
          nullable: true
          description: Camera manufacturer from EXIF
          example: "Canon"
        cameraModel:
          type: string
          nullable: true
          description: Camera model from EXIF
          example: "Canon EOS R5"
        focalLength:
          type: number
          nullable: true
          description: Focal length in mm
          example: 85
        aperture:
          type: number
          nullable: true
          description: Aperture (f-stop)
          example: 1.8
        shutterSpeed:
          type: string
          nullable: true
          description: Shutter speed as fraction string
          example: "1/250"
        iso:
          type: integer
          nullable: true
          description: ISO sensitivity
          example: 400
        takenAt:
          type: string
          format: date-time
          nullable: true
          description: Date/time photo was taken (from EXIF)
        latitude:
          type: number
          nullable: true
          description: GPS latitude
          example: 48.8566
        longitude:
          type: number
          nullable: true
          description: GPS longitude
          example: 2.3522
        isPublic:
          type: boolean
          description: Whether image is publicly accessible
          example: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Associated tags
        createdAt:
          type: string
          format: date-time
          description: Upload timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    ImageCreate:
      type: object
      description: Multipart form data for image upload
      required:
        - image
      properties:
        image:
          type: string
          format: binary
          description: Image file (JPEG, PNG, GIF, WebP, TIFF, HEIC)
        title:
          type: string
          description: Optional title for the image
        description:
          type: string
          description: Optional description
        tags:
          type: string
          description: Comma-separated list of tag names
    ImageUpdate:
      type: object
      description: Fields to update on an image
      properties:
        title:
          type: string
          nullable: true
          example: "Updated title"
        description:
          type: string
          nullable: true
          example: "Updated description"
        altText:
          type: string
          nullable: true
          example: "Updated alt text"
        isPublic:
          type: boolean
          example: true
        tags:
          type: array
          items:
            type: string
          description: Replace all tags with this list
          example: ["nature", "sunset", "beach"]
    Tag:
      type: object
      properties:
        id:
          type: integer
          description: Unique tag identifier
          example: 7
        name:
          type: string
          description: Tag name (lowercase)
          example: "nature"
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total:
          type: integer
          description: Total number of items
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 8
        hasNext:
          type: boolean
          description: Whether a next page exists
          example: true
        hasPrev:
          type: boolean
          description: Whether a previous page exists
          example: false

    # ─── Errors ──────────────────────────────────────────────────────────────────
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
              example: "NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Image with ID 42 not found"

    # ─── Search ──────────────────────────────────────────────────────────────────
...
    AdminStats:
      type: object
      properties:
        totals:
          type: object
          properties:
            images:
              type: integer
              example: 1250
            storageBytes:
              type: integer
              example: 5368709120
            storageMB:
              type: number
              example: 5120.5
            tags:
              type: integer
              example: 87
            apiKeys:
              type: integer
              example: 3
        topTags:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              count:
                type: integer
        cameraMakes:
          type: array
          items:
            type: object
            properties:
              make:
                type: string
              count:
                type: integer
        uploadsPerMonth:
          type: array
          items:
            type: object
            properties:
              month:
                type: string
                example: "2026-06"
              count:
                type: integer
                example: 42

# ─── Security (Global) ────────────────────────────────────────────────────────
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/images:
    get:
      operationId: listImages
      summary: List images
      description: Returns a paginated list of images with optional filtering and sorting.
      tags: [Images]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: search
          in: query
          schema:
            type: string
          description: Search by filename or description
        - name: cameraMake
          in: query
          schema:
            type: string
          description: Filter by camera manufacturer
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, fileSize, originalName]
            default: createdAt
          description: Sort field
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: uploadImage
      summary: Upload an image
      description: Upload a new image file with optional metadata.
      tags: [Images]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ImageCreate'
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Image'
        '400':
          description: Bad request - no image file or invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/images/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Image ID

    get:
      operationId: getImage
      summary: Get an image
      description: Get a single image by ID, including its tags.
      tags: [Images]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Image'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateImage
      summary: Update image metadata
      description: Update title, description, alt text, visibility, or tags.
      tags: [Images]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageUpdate'
      responses:
        '200':
          description: Image updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Image'
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteImage
      summary: Delete an image
      description: Permanently delete an image and its associated files.
      tags: [Images]
      responses:
        '204':
          description: Image deleted successfully
        '404':
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Tags ────────────────────────────────────────────────────────────────────
...
  /admin/stats:
    get:
      operationId: getAdminStats
      summary: Get admin statistics
      description: |
        Returns aggregate statistics: total images, storage, tags, API keys,
        top tags, camera make distribution, and uploads per month.
        
        **Authentication**: HTTP Basic Auth (ADMIN_USER / ADMIN_PASS env vars)
      tags: [Admin]
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Statistics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/AdminStats'
        '401':
          description: Unauthorized
  /admin/cache/flush:
    post:
      operationId: flushCache
      summary: Flush cache
      description: Flush the thumbnail and response cache.
      tags: [Admin]
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Cache flushed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized

  # ── Docs ─────────────────────────────────────────────────────────────────────
...